Remove star import from network.py.
Also move some methods only used by network.py from util.py.
